Adjusted some things in section 2
Before my changes, it kept referencing the Arithmetic calculator project from last semester. I updated it to refer to the Boolean logic calculator we are making this semester.
I also reworded a couple of lines that sounded confusing to me.
Finally, I expanded the constraints section 2.4 so that the requirements from the project description are actually included in this document. (Before someone reading the SRS would have to refer to the project description document, now they should have all of the information in one place.)